refactor: refactor the controller to include responseentity in my routes and customize responses to client requests

+++ src/main/java/dev/joaofvic/NinjaRegistry/Ninjas/NinjaController.java
@@ -1,5 +1,7 @@
 package dev.joaofvic.NinjaRegistry.Ninjas;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -25,32 +27,54 @@ public class NinjaController {
 
     // Add ninja
     @PostMapping("/register")
-    public NinjaDTO register(@RequestBody NinjaDTO ninjaDTO) {
-        return ninjaService.register(ninjaDTO);
+    public ResponseEntity<String> register(@RequestBody NinjaDTO ninjaDTO) {
+        NinjaDTO ninjaNew = ninjaService.register(ninjaDTO);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Ninja created successfully: " + ninjaNew.getName());
     }
 
     // Show ninjas
     @GetMapping("/all")
-    public List<NinjaDTO> all() {
-        return ninjaService.listNinjas();
+    public ResponseEntity<List<NinjaDTO>> all() {
+        List<NinjaDTO> ninjas = ninjaService.listNinjas();
+        return ResponseEntity.ok(ninjas);
     }
 
     // Show ninjas by Id
     @GetMapping("/all/{id}")
-    public NinjaDTO allId(@PathVariable Long id) {
-        return ninjaService.listNinjasId(id);
+    public ResponseEntity<?> allId(@PathVariable Long id) {
+        NinjaDTO ninja = ninjaService.listNinjasId(id);
+
+        if (ninja != null) {
+            return ResponseEntity.ok(ninja);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja from id " + id + " not found.");
+        }
     }
 
     // Change ninja data
-    @PutMapping("/changeId/{id}")
-    public NinjaDTO changeId(@PathVariable Long id, @RequestBody NinjaDTO ninjaUpdated) {
-        return ninjaService.changeId(id, ninjaUpdated);
+    @PutMapping("/change/{id}")
+    public ResponseEntity<String> changeId(@PathVariable Long id, @RequestBody NinjaDTO ninjaUpdated) {
+        if (ninjaService.listNinjasId(id) != null) {
+            ninjaService.changeId(id, ninjaUpdated);
+            return ResponseEntity.ok().body("Ninja changed successfully");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja from id " + id + " not found");
+        }
     }
 
     // Delete ninja
     @DeleteMapping("/deleteId/{id}")
-    public void deleteId(@PathVariable Long id) {
-        ninjaService.deleteId(id);
+    public ResponseEntity<String> deleteId(@PathVariable Long id) {
+
+        if (ninjaService.listNinjasId(id) != null) {
+            ninjaService.deleteId(id);
+            return ResponseEntity.ok("Ninja from id " + id + " deleted");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja from id " + id + " not found");
+        }
     }
 
 }
